Added a draft of the following:
Change the text color or background of a section when hovered over or clicked.
Create a button that, when clicked, displays a fun fact about yourself or a random quote.
Toggle Visibility: Allow sections of the CV (like work experience or skills) to expand or collapse when clicked.

Still have to add the following:

Contact US
- Check for a valid email format.
- Ensure required fields are filled.
- Display error messages when validation fails (alert).

General
- Theme Switcher: Implement a button that lets users toggle between light and dark modes for their CV.